% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pages.R
\name{build_data_page}
\alias{build_data_page}
\title{Build a summary page for a data set}
\usage{
build_data_page(di, title = "", author = "", output_dir, ids = NULL,
  view = FALSE)
}
\arguments{
\item{di}{an object returned from \code{\link{get_data_info}} or data frame}

\item{title}{page title}

\item{author}{name of author}

\item{output_dir}{directory where output html file should be placed}

\item{ids}{an optional list of ids to be used to generate links to pages for other data sets}

\item{view}{should the page be opened in a browser after it is rendered?}
}
\description{
Build a summary page for a data set
}

