% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pages.R
\name{build_index_page}
\alias{build_index_page}
\title{Build index pages for a set of data sets}
\usage{
build_index_page(di_list, title = "", author = "", output_dir,
  intro_rmd = "", ...)
}
\arguments{
\item{di_list}{a list of objects returned from \code{\link{get_data_info}} or named list of data frames}

\item{title}{page title}

\item{author}{name of author}

\item{output_dir}{directory where output html files should be placed}

\item{intro_rmd}{optional R markdown string that will be placed at beginning of the page}

\item{\ldots}{data to be saved and loaded in to the document's knitr environment}
}
\description{
Build index pages for a set of data sets
}

