% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pages.R
\name{build_pages}
\alias{build_pages}
\title{Build summary pages for a set of data sets with an optional index page}
\usage{
build_pages(di_list, title = "", author = "", output_dir, index = TRUE,
  intro_rmd = "", view = FALSE)
}
\arguments{
\item{di_list}{a list of objects returned from \code{\link{get_data_info}} or named list of data frames}

\item{title}{page title}

\item{author}{name of author}

\item{output_dir}{directory where output html files should be placed}

\item{index}{should an index page be built as well?}

\item{intro_rmd}{passed on to \code{\link{build_index_page}}}

\item{view}{should the index page be opened in a browser after it is rendered?}
}
\description{
Build summary pages for a set of data sets with an optional index page
}

