% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\title{Summary of Numeric Columns}
\usage{
data_summary(x, na.rm = TRUE)
}
\arguments{
\item{x}{A data frame. Non-numeric columns will be removed}

\item{na.rm}{A logical indicating whether missing values should be removed}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
\item ID
\item min
\item median
\item sd
\item max
}
}
\description{
Generate specific summaries of numeric columns in a data frame
}
\examples{
data_summary(iris)
data_summary(airquality, na.rm = FALSE)
\dontrun{
data_summary(airquality, na.rm = FALSE)
}
}
\seealso{
\link[base]{summary}
}
\author{
Jean-Michel Bodart \href{mailto:jean-michel.bodart@businessdecision.com}{jean-michel.bodart@businessdecision.com}
}
