% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary.R
\docType{package}
\name{datasummary}
\alias{datasummary}
\alias{datasummary-package}
\title{Custom Data Summaries}
\description{
Easily generate custom data frame summaries
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmbo1190/datasummary}
  \item Report bugs at \url{https://github.com/jmbo1190/datasummary/issues}
}

}
\author{
\strong{Maintainer}: Jean-Michel Bodart \email{Jean-Michel.Bodart@BusinessDecision.com} (\href{https://orcid.org/0000-0002-9957-6653}{ORCID})

}
