% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{get_data_info}
\alias{get_data_info}
\title{Get all summary data necessary for generatng a summary report}
\usage{
get_data_info(dat, id = digest::digest(dat), ...)
}
\arguments{
\item{dat}{a data frame}

\item{id}{optional id to use}

\item{\ldots}{optional additional attributes to store with the returned object (such as additional meta data)}
}
\description{
Get all summary data necessary for generatng a summary report
}

