% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{group_summary}
\alias{group_summary}
\title{Summarise the dataset by group}
\usage{
group_summary(data, ...)
}
\arguments{
\item{data}{dataset to summarise}

\item{...}{ellipses, column(s) to group by}
}
\value{
the summary result
}
\description{
For each group, calculate the number of rows, numeric columns' mean value
and categorical columns' number of unique values
}
\examples{
group_summary(penguins, island)
group_summary(penguins, species, island)

}
