% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_summary.R
\name{numeric_summary}
\alias{numeric_summary}
\title{Numeric Summaries
Summarises numeric data and returns a data frame containing the minimum value, median, standard deviation, and maximum value.}
\usage{
numeric_summary(x, na.rm)
}
\arguments{
\item{x}{a numeric vector containing the values to summarize.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
Numeric Summaries
Summarises numeric data and returns a data frame containing the minimum value, median, standard deviation, and maximum value.
}
