% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable.R
\name{plot_cat}
\alias{plot_cat}
\title{Plot a categorical frequency table}
\usage{
plot_cat(dtab, width = 750, min_height = 400)
}
\arguments{
\item{dtab}{a frequency table data frame}

\item{width}{width of the plot in pixels}

\item{min_height}{minimum height of the plot in pixels (the actual height is computed as a function of the number of rows in \code{dtab})}
}
\description{
Plot a categorical frequency table
}

