% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intra.R
\name{plot_common_vars_freq}
\alias{plot_common_vars_freq}
\title{Plot frequency of common variables across data sets}
\usage{
plot_common_vars_freq(di_list)
}
\arguments{
\item{di_list}{a list of objects returned from \code{\link{get_data_info}}}
}
\description{
Plot frequency of common variables across data sets
}

