% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intra.R
\name{plot_common_vars_tab}
\alias{plot_common_vars_tab}
\title{Plot dataset / variable pairs}
\usage{
plot_common_vars_tab(di_list)
}
\arguments{
\item{di_list}{a list of objects returned from \code{\link{get_data_info}}}
}
\description{
Plot dataset / variable pairs
}

