% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable.R
\name{plot_missing}
\alias{plot_missing}
\title{Plot number of missing values for each variable in a data frame}
\usage{
plot_missing(di, width = 800, height = 600, ...)
}
\arguments{
\item{di}{an object returned from \code{\link{get_data_info}}}

\item{width}{width of the plot in pixels}

\item{height}{height of the plot in pixels}

\item{\ldots}{additional parameters passed to \code{\link[rbokeh]{figure}}}
}
\description{
Plot number of missing values for each variable in a data frame
}

