% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intra.R
\name{plot_nrows}
\alias{plot_nrows}
\title{Plot number of rows in each data frame}
\usage{
plot_nrows(di_list)
}
\arguments{
\item{di_list}{a list of objects returned from \code{\link{get_data_info}}}
}
\description{
Plot number of rows in each data frame
}

