% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable.R
\name{plot_num}
\alias{plot_num}
\title{Plot a histogram for a numeric variable}
\usage{
plot_num(hst, xlab = "x", log = FALSE, width = 750, height = 520)
}
\arguments{
\item{hst}{an object returned from \code{\link[graphics]{hist}}}

\item{xlab}{x-axis label}

\item{log}{should the x-axis be shown on the log scale?}

\item{width}{width of the plot in pixels}

\item{height}{height of the plot in pixels}
}
\description{
Plot a histogram for a numeric variable
}

