% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\title{Sumario de Colunas Numericas}
\description{
Gera um sumario das colunas numericas de um data.frame
}
\usage{
data_summary(x, na.rm = TRUE)
}
\arguments{
\item{x}{Um data frame com colunas numericas}

\item{na.rm}{Logico, indicando se NAs devem ser removidos}
}
\value{
Essa funcao retorna um \code{data.frame} com as colunas:
\itemize{
\item ID
\item min
\item median
\item sd
\item max
}
}
\description{
Sumario de Colunas Numericas
Gera um sumario das colunas numericas de um data.frame
}
\examples{
data_summary(iris)
data_summary(airquality, na.rm = FALSE)

}
\seealso{
\link[base]{summary}
}
\author{
Elton J (elton.pdsj@gmail.com)
}
