% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary.R
\docType{package}
\name{datasummary}
\alias{datasummary}
\alias{datasummary-package}
\title{datasummary: Sumarios Numericos}
\description{
Gera sumarios de data.frame e vetores
}
\author{
\strong{Maintainer}: Elton Junior \email{elton.pdsj@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
