% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_summary.R
\name{numeric_summary}
\alias{numeric_summary}
\title{Sumario de Vetor Numerico}
\usage{
numeric_summary(x, na.rm = T)
}
\arguments{
\item{x}{Um vetor numerico}

\item{na.rm}{Logico, indicando se NAs devem ser removidos}
}
\value{
Essa funcao retorna um \code{data.frame} com as colunas:
\itemize{
\item min
\item median
\item sd
\item max
}
}
\description{
Gera um sumario de um vetor numerico
}
\examples{
numeric_summary(c(1:10))
}
\seealso{
\link[base]{summary}
}
\author{
Elton J (elton.pdsj@gmail.com)
}
